# Change mesh vertices and indices to accept memoryviews.