@tito 2016-09-20
I'm fixing and getting back to work this PR.